%     The AHIsmooth R package
%     Copyright (C) 2016 Wesley Burr
%
%     Written by Wesley Burr.
%
%     This file is part of the AHItools ecosystem.
%
%     The AHIsmooth package is free software: you can redistribute it and
%     or modify it under the terms of the GNU General Public License as 
%     published by the Free Software Foundation, either version 2 of the 
%     License, or any later version.
%
%     The AHIsmooth package is distributed in the hope that it will be 
%     useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
%     of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%     GNU General Public License for more details.
%
%     You should have received a copy of the GNU General Public License
%     along with AHIsmooth. If not, see <http://www.gnu.org/licenses/>.
%
%     If you wish to report bugs please contact the author. 
%     Wesley Burr
%     wesley.burr@gmail.com


\name{AHIsmooth-package}
\alias{AHIsmooth-package}
\alias{AHIsmooth}
\docType{package}
\title{
    AHIsmooth: SLP Time Smoothers for Estimation of the Air Health Indicator
}
\description{
    SLP time smoother objects for the AHItools package.

}
\details{
\tabular{ll}{
Package: \tab AHIsmooth\cr
Type: \tab Package\cr
Version: \tab 2016.02-0\cr
Date: \tab 2016-02-19\cr
License: \tab GPL-2 \cr
}
}
\author{
Wesley Burr <wesley.burr@gmail.com>
}
\references{
}
\keyword{ }
\seealso{
}
\examples{
}
